@@ -3,7 +3,7 @@ import Todos from "../models/todos.model.js";
 
 // add todo
 
-const addTodo = (req, res) => {
+export const addTodo = (req, res) => {
   const { title, description } = req.body;
 
   if (!title || !description) {
@@ -21,4 +21,39 @@ const addTodo = (req, res) => {
   });
 };
 
-export { addTodo };
+// To get all todos
+
+export const getallTodos = (req , res) =>{
+  if(!Todos){
+    res.status(400).json({
+      message:"No Todos added Yet",
+      Todos,
+    })
+    return;
+  }
+  res.status(200).json({
+    message:"All Todos ==>",
+    Todos,
+  })
+}
+
+
+
+// To get all single todo
+
+export const getsingleTodo = (req , res) =>{
+  const { id } = req.params;
+   
+  
+  if(!Todos){
+    res.status(400).json({
+      message:"No Todos added Yet",
+      Todos,
+    })
+    return;
+  }
+  res.status(200).json({
+    message:"All Todos ==>",
+    Todos,
+  })
+}
